+++ src/plugins/libraries/vue-carousel.ts
