+++ src/plugins/libraries/vue-lazyload.ts
