+++ src/plugins/locale/i18n.ts
@@ -14,7 +14,7 @@ export default ({ app, store }): void => {
     locale: store.state.i18n.locale,
     fallbackLocale: 'ja',
     messages: {
-      ja: require('~/locales/ja.json')
+      ja: require('@/locales/ja.json')
     }
   })
